o11y-playground/o11y-python/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
